+++ samples/Demo/Beef.Demo.Business/Data/ZippoAgent.cs
@@ -8,9 +8,6 @@ namespace Beef.Demo.Business.Data
     public class ZippoAgent : TypedMappedHttpClientCore<ZippoAgent>
     {
         public ZippoAgent(HttpClient client, IMapper mapper, IJsonSerializer jsonSerializer, CoreEx.ExecutionContext executionContext, SettingsBase settings, ILogger<ZippoAgent> logger)
-            : base(client, mapper, jsonSerializer, executionContext, settings, logger)
-        {
-            DefaultOptions.WithRetry();
-        }
+            : base(client, mapper, jsonSerializer, executionContext, settings, logger) { }
     }
 }