+++ samples/MyEf.Hr/MyEf.Hr.Security.Subscriptions/GlobalUsings.cs
@@ -15,6 +15,7 @@ global using Microsoft.Extensions.Configuration;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.Hosting;
 global using Microsoft.Extensions.Logging;
+global using Polly;
 global using System;
 global using System.Collections.Generic;
 global using System.Linq;