+++ tools/Beef.CodeGen.Core/Config/Entity/OperationConfig.cs
@@ -525,6 +525,14 @@ operations: [
             Description = "Defaults to the _Common_ type. A value of `None`, `none` or `` will ensure no type is emitted.")]
         public string? WebApiProducesResponseType { get; set; }
 
+        /// <summary>
+        /// The list of tags to add for the generated `WebApi`.
+        /// </summary>
+        [JsonPropertyName("webApiTags")]
+        [CodeGenPropertyCollection("WebApi", Title = "The list of tags to add for the generated `WebApi` operation.",
+            Description = "Overrides the `Entity.WebApiTags`; unless, if the first tag value is a `^` then this indicates that the `Entity.WebApiTags` are to be included (inherited) as a replacement. Otherwise, defaults to `Entity.WebApiTags`.")]
+        public List<string>? WebApiTags { get; set; }
+
         #endregion
 
         #region Auth
@@ -992,6 +1000,11 @@ operations: [
         public string? WebApiProducesContentType => WebApiProduces is not null && WebApiProduces.Count > 0 ? string.Join(", ", WebApiProduces.Select(x => $"\"{x}\"")) : null;
 
         /// <summary>
+        /// Gets the Tags for reporting.
+        /// </summary>
+        public List<string> ReportTags => WebApiTags!.Count >= 1 ? WebApiTags : Parent!.WebApiTags!;
+
+        /// <summary>
         /// <inheritdoc/>
         /// </summary>
         protected override async Task PrepareAsync()
@@ -1184,6 +1197,23 @@ operations: [
             ExcludeWebApi = DefaultWhereNull(ExcludeWebApi, () => CompareValue(ExcludeAll, true));
             ExcludeWebApiAgent = DefaultWhereNull(ExcludeWebApiAgent, () => CompareValue(ExcludeAll, true));
             ExcludeGrpcAgent = DefaultWhereNull(ExcludeGrpcAgent, () => CompareValue(ExcludeAll, true));
+
+            // Ensure the WebApiTags are set correctly.
+            if (WebApiTags is not null && WebApiTags.Any())
+            {
+                var list = new List<string>();
+                foreach (var tag in WebApiTags)
+                {
+                    if (tag == "^")
+                        list.AddRange(Parent!.WebApiTags!);
+                    else
+                        list.Add(tag);
+                }
+
+                WebApiTags = list;
+            }
+            else
+                WebApiTags ??= [];
 
             if (Type == "Patch")
                 ExcludeIData = ExcludeData = ExcludeIDataSvc = ExcludeDataSvc = ExcludeIManager = ExcludeManager = true;